test(popover): migrate trigger
